Refactor of interface names
Interface names refactored so they are have the same convention with the project
