+++ mrchecker-framework-modules/mrchecker-core-module/src/main/java/com/capgemini/mrchecker/test/core/utils/encryption/controller/IDataEncryptionController.java
@@ -1,0 +1,13 @@
+package com.capgemini.mrchecker.test.core.utils.encryption.controller;
+
+import com.capgemini.mrchecker.test.core.utils.encryption.CryptParams;
+import com.capgemini.mrchecker.test.core.utils.encryption.view.IDataEncryptionView;
+
+public interface IDataEncryptionController {
+	
+	void onEncrypt(CryptParams cryptParams);
+	
+	void onDecrypt(CryptParams cryptParams);
+	
+	void setView(IDataEncryptionView view);
+}
