+++ mrchecker-framework-modules/mrchecker-core-module/src/main/java/com/capgemini/mrchecker/test/core/utils/encryption/view/IDataEncryptionView.java
@@ -1,6 +1,6 @@
 package com.capgemini.mrchecker.test.core.utils.encryption.view;
 
-public interface DataEncryptionView {
+public interface IDataEncryptionView {
 	void setEncryptionFieldValue(String text);
 	
 	void setDecryptionFieldValue(String text);
